renamed controllers to views
